Can add new client but without validation
Major change in the database. Can already insert new client but failed to do validation. The errorBags failed to contain the errors during validation.

The action column contains an edit button but it only retrieve the personal data of the client from the clients table.

The datatable failed to render Vue components in the action column. Still finding a way to make it work.